Bug 1969124 - Part 2: Cleanup toolbar shortcut preferences r=android-reviewers,Roger
Differential Revision: https://phabricator.services.mozilla.com/D271496
