+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/settings/ToolbarShortcut.kt
@@ -6,6 +6,8 @@ package org.mozilla.fenix.settings
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
+import org.mozilla.fenix.R
+import mozilla.components.ui.icons.R as iconsR
 
 /**
  * String keys used to persist and map the selected toolbar shortcut option.
@@ -20,8 +22,59 @@ object ShortcutType {
     const val BACK = "back"
 }
 
+internal enum class ShortcutAvailability {
+    SIMPLE,
+    EXPANDED,
+}
+
 internal data class ShortcutOption(
     val key: String,
     @param:DrawableRes val icon: Int,
     @param:StringRes val label: Int,
+    val availability: Set<ShortcutAvailability>,
 )
+
+internal val allShortcutOptions: List<ShortcutOption> = listOf(
+    ShortcutOption(
+        ShortcutType.NEW_TAB,
+        iconsR.drawable.mozac_ic_plus_24,
+        R.string.toolbar_customize_shortcut_new_tab,
+        setOf(ShortcutAvailability.SIMPLE),
+    ),
+    ShortcutOption(
+        ShortcutType.SHARE,
+        iconsR.drawable.mozac_ic_share_android_24,
+        R.string.toolbar_customize_shortcut_share,
+        setOf(ShortcutAvailability.SIMPLE),
+    ),
+    ShortcutOption(
+        ShortcutType.BOOKMARK,
+        iconsR.drawable.mozac_ic_bookmark_24,
+        R.string.toolbar_customize_shortcut_add_bookmark,
+        setOf(ShortcutAvailability.SIMPLE, ShortcutAvailability.EXPANDED),
+    ),
+    ShortcutOption(
+        ShortcutType.TRANSLATE,
+        iconsR.drawable.mozac_ic_translate_24,
+        R.string.toolbar_customize_shortcut_translate,
+        setOf(ShortcutAvailability.SIMPLE),
+    ),
+    ShortcutOption(
+        ShortcutType.HOMEPAGE,
+        iconsR.drawable.mozac_ic_home_24,
+        R.string.toolbar_customize_shortcut_homepage,
+        setOf(ShortcutAvailability.SIMPLE),
+    ),
+    ShortcutOption(
+        ShortcutType.BACK,
+        iconsR.drawable.mozac_ic_back_24,
+        R.string.toolbar_customize_shortcut_back,
+        setOf(ShortcutAvailability.SIMPLE),
+    ),
+)
+
+internal val expandedShortcutOptions: List<ShortcutOption> =
+    allShortcutOptions.filter { ShortcutAvailability.EXPANDED in it.availability }
+
+internal val simpleShortcutOptions: List<ShortcutOption> =
+    allShortcutOptions.filter { ShortcutAvailability.SIMPLE in it.availability }
